Add workaround to handle Commit: diff views
IntelliJ expect the base revision for files to be HEAD whenever a
Commit: diff is opened. This is not the case when a git scope other
than HEAD is selected. This commit introduces a workaround where
gitscope will monitor the active and opened tab, and temporarly
restores the scope to HEAD for files active in a Commit: diff view.
As soon as it is not in focus or closed the correct scope is restored.

Fixes comod/git-scope-pro#56